requirements.txt
discord.py
dotenv
responses